+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/controller/model/v1/FileMetadataDto.java
@@ -1,18 +1,17 @@
-package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model;
+package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v1;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
-import it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.model.AggregatesDataSummary;
+import java.time.LocalDateTime;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import java.time.LocalDateTime;
 import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
 
 @Data
-public class FileMetadataV2Dto {
+public class FileMetadataDto {
 
   @NotNull
   @NotBlank
@@ -29,6 +28,4 @@ public class FileMetadataV2Dto {
   @JsonDeserialize(using = LocalDateTimeDeserializer.class)
   private LocalDateTime transmissionDate;
 
-  @NotNull
-  private AggregatesDataSummary dataSummary;
 }
