+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/controller/model/v1/FileReportDto.java
@@ -1,0 +1,10 @@
+package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v1;
+
+import java.util.Collection;
+import lombok.Data;
+
+@Data
+public class FileReportDto {
+
+  Collection<FileMetadataDto> filesRecentlyUploaded;
+}
