+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/controller/model/v1/FileReportDtoMapper.java
@@ -1,5 +1,6 @@
-package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model;
+package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v1;
 
+import it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v1.FileReportDto;
 import it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.model.FileReport;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
