+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/controller/model/v2/FileMetadataDto.java
@@ -1,12 +1,13 @@
-package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model;
+package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v2;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
-import java.time.LocalDateTime;
+import it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.model.AggregatesDataSummary;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import java.time.LocalDateTime;
 import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
 
@@ -28,4 +29,6 @@ public class FileMetadataDto {
   @JsonDeserialize(using = LocalDateTimeDeserializer.class)
   private LocalDateTime transmissionDate;
 
+  @NotNull
+  private AggregatesDataSummary dataSummary;
 }
