+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/controller/model/v2/FileReportV2DtoMapper.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model;
+package it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v2;
 
 import it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.model.FileMetadata;
 import it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.model.FileReport;
@@ -9,8 +9,8 @@ import org.mapstruct.Mapping;
 public interface FileReportV2DtoMapper {
 
   @Mapping(source = "filesUploaded", target = "filesRecentlyUploaded")
-  FileReportV2Dto fileReportToDto(FileReport fileReport);
+  FileReportDto fileReportToDto(FileReport fileReport);
 
   @Mapping(source = "aggregatesDataSummary", target = "dataSummary")
-  FileMetadataV2Dto fileMetadataToDto(FileMetadata fileMetadata);
+  FileMetadataDto fileMetadataToDto(FileMetadata fileMetadata);
 }
